move `_SourceFactories` to `sources.py` module
